[BBB-129] 불필요 코드 제거 및 이동
