+++ src/main/java/com/bangbangbwa/backend/domain/sns/controller/SnsApi.java
@@ -69,62 +69,62 @@ public interface SnsApi {
   )
   public ApiResponse<GetPostDetailsDto.Response> getPostDetails(@Parameter Long postId);
 
-
-  @Operation(
-      tags = {"SnsAPI"},
-      summary = "게시글 작성",
-      description = "게시글을 작성합니다."
-  )
-  @ApiResponses(value = {
-      @io.swagger.v3.oas.annotations.responses.ApiResponse(
-          responseCode = "200",
-          description = "OK",
-          content = @Content(
-              mediaType = MediaType.APPLICATION_JSON_VALUE,
-              schema = @Schema(implementation = CreatePostDto.Response.class)
-          )
-      ),
-      @io.swagger.v3.oas.annotations.responses.ApiResponse(
-          responseCode = "401",
-          description = "권한 없음",
-          content = @Content(
-              mediaType = "application/json",
-              schema = @Schema(implementation = ApiResponse.class),
-              examples = @ExampleObject(
-                  value = """
-                      {
-                      "code": "NO_POST_PERMISSION",
-                      "message": "작성 권한이 없습니다.",
-                      "data" : null
-                      }
-                      """
-              )
-          )
-      ),
-      @io.swagger.v3.oas.annotations.responses.ApiResponse(
-          responseCode = "400",
-          description = "잘못된 요청",
-          content = @Content(
-              mediaType = "application/json",
-              schema = @Schema(implementation = ApiResponse.class),
-              examples = @ExampleObject(
-                  value = """
-                      {
-                      "code": "BAD_REQUEST",
-                      "message": "잘못된 요청입니다.",
-                      "data" : [
-                        "postType: 게시글 타입을 입력해주세요.",
-                        "title: 제목을 입력해주세요.",
-                        "content: 내용을 입력해주세요."
-                      ]
-                      }
-                      """
-              )
-          )
-      )
-  })
-  ApiResponse<CreatePostDto.Response> createPost(
-      @RequestBody CreatePostDto.Request request
-  );
+//
+//  @Operation(
+//      tags = {"SnsAPI"},
+//      summary = "게시글 작성",
+//      description = "게시글을 작성합니다."
+//  )
+//  @ApiResponses(value = {
+//      @io.swagger.v3.oas.annotations.responses.ApiResponse(
+//          responseCode = "200",
+//          description = "OK",
+//          content = @Content(
+//              mediaType = MediaType.APPLICATION_JSON_VALUE,
+//              schema = @Schema(implementation = CreatePostDto.Response.class)
+//          )
+//      ),
+//      @io.swagger.v3.oas.annotations.responses.ApiResponse(
+//          responseCode = "401",
+//          description = "권한 없음",
+//          content = @Content(
+//              mediaType = "application/json",
+//              schema = @Schema(implementation = ApiResponse.class),
+//              examples = @ExampleObject(
+//                  value = """
+//                      {
+//                      "code": "NO_POST_PERMISSION",
+//                      "message": "작성 권한이 없습니다.",
+//                      "data" : null
+//                      }
+//                      """
+//              )
+//          )
+//      ),
+//      @io.swagger.v3.oas.annotations.responses.ApiResponse(
+//          responseCode = "400",
+//          description = "잘못된 요청",
+//          content = @Content(
+//              mediaType = "application/json",
+//              schema = @Schema(implementation = ApiResponse.class),
+//              examples = @ExampleObject(
+//                  value = """
+//                      {
+//                      "code": "BAD_REQUEST",
+//                      "message": "잘못된 요청입니다.",
+//                      "data" : [
+//                        "postType: 게시글 타입을 입력해주세요.",
+//                        "title: 제목을 입력해주세요.",
+//                        "content: 내용을 입력해주세요."
+//                      ]
+//                      }
+//                      """
+//              )
+//          )
+//      )
+//  })
+//  ApiResponse<CreatePostDto.Response> createPost(
+//      @RequestBody CreatePostDto.Request request
+//  );
 
 }
