+++ src/main/java/com/bangbangbwa/backend/domain/sns/controller/SnsController.java
@@ -47,15 +47,15 @@ public class SnsController implements SnsApi {
     return ApiResponse.ok(response);
   }
 
-  @PostMapping("/createPost")
-  @PreAuthorize("hasAuthority('MEMBER')")
-  public ApiResponse<CreatePostDto.Response> createPost(
-      @RequestBody @Valid CreatePostDto.Request request
-  ) {
-    Post post = snsService.createPost(request);
-    CreatePostDto.Response response = PostMapper.INSTANCE.dtoToCreatePostResponse(post);
-    return ApiResponse.ok(response);
-  }
+//  @PostMapping("/createPost")
+//  @PreAuthorize("hasAuthority('MEMBER')")
+//  public ApiResponse<CreatePostDto.Response> createPost(
+//      @RequestBody @Valid CreatePostDto.Request request
+//  ) {
+//    Post post = snsService.createPost(request);
+//    CreatePostDto.Response response = PostMapper.INSTANCE.dtoToCreatePostResponse(post);
+//    return ApiResponse.ok(response);
+//  }
 
   @PostMapping(value = "/uploadPostMedia", consumes = {MediaType.MULTIPART_FORM_DATA_VALUE})
   @PreAuthorize("hasAuthority('MEMBER')")
