+++ src/test/java/com/bangbangbwa/backend/domain/sns/controller/SnsIntegrationTest.java
@@ -345,132 +345,6 @@ class SnsIntegrationTest extends IntegrationTest {
 
   }
 
-
-  // note. PostType 랜덤 지정하도록 변경 필요
-  @Test
-  void createPost_성공() {
-    // given
-    Member member = createMember();
-    TokenDto tokenDto = tokenProvider.getToken(member);
-
-    List<Long> nullList = new ArrayList<>();
-    CreatePostDto.Request request = new CreatePostDto.Request(
-        PostType.MEMBER,
-        RandomValue.string(100).setNullable(false).get(),
-        RandomValue.string(2000).setNullable(false).get(),
-        nullList,
-        nullList
-    );
-
-    String url = "http://localhost:" + port + "/api/v1/sns/createPost";
-
-    HttpHeaders headers = new HttpHeaders();
-    headers.setBearerAuth(tokenDto.getAccessToken());
-    HttpEntity<CreatePostDto.Request> requestEntity = new HttpEntity<>(request, headers);
-
-    // when
-    ResponseEntity<String> responseEntity = restTemplate.postForEntity(
-        url,
-        requestEntity,
-        String.class
-    );
-
-    ApiResponse<CreatePostDto.Response> apiResponse = gson.fromJson(
-        responseEntity.getBody(),
-        new TypeToken<ApiResponse<CreatePostDto.Response>>() {
-        }.getType()
-    );
-
-    // then
-    assertThat(responseEntity.getStatusCode()).isEqualTo(HttpStatus.OK);
-    assertNotNull(apiResponse.getData());
-    assertThat(apiResponse.getData().postType()).isEqualTo(request.postType());
-    assertThat(apiResponse.getData().title()).isEqualTo(request.title());
-    assertThat(apiResponse.getData().content()).isEqualTo(request.content());
-
-  }
-
-
-  @Test
-  void createPost_공개_비공개_중복_입력() {
-    // given
-    Member member = createMember();
-    TokenDto tokenDto = tokenProvider.getToken(member);
-
-    List<Long> memberIds = List.of(1L, 2L);
-    CreatePostDto.Request request = new CreatePostDto.Request(
-        PostType.MEMBER,
-        RandomValue.string(100).setNullable(false).get(),
-        RandomValue.string(2000).setNullable(false).get(),
-        memberIds,
-        memberIds
-    );
-
-    String url = "http://localhost:" + port + "/api/v1/sns/createPost";
-
-    InvalidMemberVisibilityException exception = new InvalidMemberVisibilityException();
-
-    HttpHeaders headers = new HttpHeaders();
-    headers.setBearerAuth(tokenDto.getAccessToken());
-    HttpEntity<CreatePostDto.Request> requestEntity = new HttpEntity<>(request, headers);
-
-    // when
-    ResponseEntity<String> responseEntity = restTemplate.postForEntity(
-        url,
-        requestEntity,
-        String.class
-    );
-
-    ApiResponse<CreatePostDto.Response> apiResponse = gson.fromJson(
-        responseEntity.getBody(),
-        new TypeToken<ApiResponse<CreatePostDto.Response>>() {
-        }.getType()
-    );
-
-    // then
-    assertThat(responseEntity.getStatusCode()).isEqualTo(HttpStatus.BAD_REQUEST);
-    assertThat(apiResponse.getCode()).isEqualTo(exception.getCode());
-    assertThat(apiResponse.getMessage()).isEqualTo(exception.getMessage());
-    assertNull(apiResponse.getData());
-  }
-
-  @Test
-  void createPost_토큰_미입력() {
-    // given
-    List<Long> memberIds = List.of(1L, 2L);
-    CreatePostDto.Request request = new CreatePostDto.Request(
-        PostType.MEMBER,
-        RandomValue.string(100).setNullable(false).get(),
-        RandomValue.string(2000).setNullable(false).get(),
-        memberIds,
-        memberIds
-    );
-
-    String url = "http://localhost:" + port + "/api/v1/sns/createPost";
-
-    AuthenticationRequiredException exception = new AuthenticationRequiredException();
-
-    // when
-    ResponseEntity<String> responseEntity = restTemplate.postForEntity(
-        url,
-        request,
-        String.class
-    );
-
-    ApiResponse<CreatePostDto.Response> apiResponse = gson.fromJson(
-        responseEntity.getBody(),
-        new TypeToken<ApiResponse<CreatePostDto.Response>>() {
-        }.getType()
-    );
-
-    // then
-    assertThat(responseEntity.getStatusCode()).isEqualTo(HttpStatus.UNAUTHORIZED);
-    assertThat(apiResponse.getCode()).isEqualTo(exception.getCode());
-    assertThat(apiResponse.getMessage()).isEqualTo(exception.getMessage());
-    assertNull(apiResponse.getData());
-  }
-
-
   @Test
   void uploadPostMedia() {
     // given
